*Changed deltatime from int to float

+++ src/Systems/luaSystem.cpp
@@ -15,9 +15,9 @@
 
 sLuaSystem::sLuaSystem():
 	hasMainBeenCalled(false),
-    lState(NULL)
+	lState(NULL)
 {
-	
+
 }
 
 sLuaSystem::~sLuaSystem()
@@ -27,32 +27,32 @@ sLuaSystem::~sLuaSystem()
 
 void sLuaSystem::Init()
 {
-    lState = luaL_newstate();
-    luaL_openlibs(lState);
-    SetPackagePath();
-    
-    mEngine::Bind(lState);
-    mInput::Bind(lState);
-    mEntity::Bind(lState);
-    mUiWindow::Bind(lState);
-    
-    string path;
-    Content::CreateFilePath("main.lua", &path);
-    
-    if (luaL_loadfile(lState, path.c_str()))
-    {
-        Terminal.Error("Could not open main.lua - The program will not run correctly");
-        return;
-    }
-    
-    lua_atpanic(lState, LuaPanic);
-    
-    //try to parse main.lua
-    if (lua_pcall(lState, 0, 0, 0) != 0)
-    {
-        Terminal.LuaError(lua_tostring(lState, -1));
-        return;
-    }
+	lState = luaL_newstate();
+	luaL_openlibs(lState);
+	SetPackagePath();
+
+	mEngine::Bind(lState);
+	mInput::Bind(lState);
+	mEntity::Bind(lState);
+	mUiWindow::Bind(lState);
+
+	string path;
+	Content::CreateFilePath("main.lua", &path);
+
+	if (luaL_loadfile(lState, path.c_str()))
+	{
+		Terminal.Error("Could not open main.lua - The program will not run correctly");
+		return;
+	}
+
+	lua_atpanic(lState, LuaPanic);
+
+	//try to parse main.lua
+	if (lua_pcall(lState, 0, 0, 0) != 0)
+	{
+		Terminal.LuaError(lua_tostring(lState, -1));
+		return;
+	}
 }
 
 void sLuaSystem::Main()
@@ -65,21 +65,22 @@ void sLuaSystem::Main()
 		{
 			Terminal.LuaError(string(lua_tostring(lState, -1)));
 		}
-        lua_settop(lState, 0);
+		lua_settop(lState, 0);
 
 		hasMainBeenCalled = true;
 	}
 }
 
-void sLuaSystem::Update(int dt)
+void sLuaSystem::Update(float dt)
 {
 	lua_getglobal(lState, "Game");
 	lua_getfield(lState, -1, "update");
 	lua_pushnumber(lState, dt);
 	if (lua_pcall(lState, 1, 0, 0) != 0)
 	{
-		Terminal.LuaError(string(lua_tostring(lState, -1)));	}
-    lua_settop(lState, 0);
+		Terminal.LuaError(string(lua_tostring(lState, -1)));
+	}
+	lua_settop(lState, 0);
 }
 
 void sLuaSystem::SendReloadCallback( const string& filePath )
@@ -89,47 +90,51 @@ void sLuaSystem::SendReloadCallback( const string& filePath )
 	lua_pushlstring( lState, filePath.c_str(), filePath.size() );
 	if (lua_pcall(lState, 1, 0, 0) != 0)
 	{
-		Terminal.LuaError(string(lua_tostring(lState, -1)));	}
+		Terminal.LuaError(string(lua_tostring(lState, -1)));
+	}
 	lua_settop(lState, 0);
 }
 
 
 void sLuaSystem::Attempt(string command)
 {
-    if(luaL_dostring(lState, command.c_str()) != 0){
-        Terminal.LuaError(string(lua_tostring(lState, -1)));
-    }
+	if (luaL_dostring(lState, command.c_str()) != 0)
+	{
+		Terminal.LuaError(string(lua_tostring(lState, -1)));
+	}
 }
 
 //TODO(robin): make this use terminal logging?
 void sLuaSystem::Dump(lua_State* L)
 {
-    int i;
-    int top = lua_gettop(L);
-    
-    printf("--=[LUA STACK DUMP]=--\n");
-    printf("total in stack %d\n",top);
-    
-    for (i = 1; i <= top; i++)
-    {  /* repeat for each level */
-        int t = lua_type(L, i);
-        switch (t) {
-            case LUA_TSTRING:  /* strings */
-                printf("string: '%s'\n", lua_tostring(L, i));
-                break;
-            case LUA_TBOOLEAN:  /* booleans */
-                printf("boolean %s\n",lua_toboolean(L, i) ? "true" : "false");
-                break;
-            case LUA_TNUMBER:  /* numbers */
-                printf("number: %g\n", lua_tonumber(L, i));
-                break;
-            default:  /* other values */
-                printf("%s\n", lua_typename(L, t));
-                break;
-        }
-        printf("  ");  /* put a separator */
-    }
-    printf("----------------------\n");/* end the listing */
+	int i;
+	int top = lua_gettop(L);
+
+	printf("--=[LUA STACK DUMP]=--\n");
+	printf("total in stack %d\n", top);
+
+	for (i = 1; i <= top; i++)
+	{
+		/* repeat for each level */
+		int t = lua_type(L, i);
+		switch (t)
+		{
+			case LUA_TSTRING:  /* strings */
+				printf("string: '%s'\n", lua_tostring(L, i));
+				break;
+			case LUA_TBOOLEAN:  /* booleans */
+				printf("boolean %s\n", lua_toboolean(L, i) ? "true" : "false");
+				break;
+			case LUA_TNUMBER:  /* numbers */
+				printf("number: %g\n", lua_tonumber(L, i));
+				break;
+			default:  /* other values */
+				printf("%s\n", lua_typename(L, t));
+				break;
+		}
+		printf("  ");  /* put a separator */
+	}
+	printf("----------------------\n");/* end the listing */
 }
 
 //PRIVATE
@@ -154,11 +159,11 @@ void sLuaSystem::SetPackagePath()
 
 int sLuaSystem::LuaPanic(lua_State* L)
 {
-    Terminal.LuaError("PANIC!");
-    Terminal.LuaError(string(lua_tostring(L, -1)));
-    Terminal.Warning("The program will exit...");
-    
-    //TODO close the lua state but keep the engine running
-    exit(0);
-    return 0;
+	Terminal.LuaError("PANIC!");
+	Terminal.LuaError(string(lua_tostring(L, -1)));
+	Terminal.Warning("The program will exit...");
+
+	//TODO close the lua state but keep the engine running
+	exit(0);
+	return 0;
 }
